SectionFiller: Access right value from tuple
The get_non_optional_settings will get us a dict with tuples as values
since we have an auto implementation of that. This allows us to auto
convert settings to given types. However the user doesn't need to know
that and he certainly doesn't want to see a tuple so lets just show the
description here.