parsing: Streamline CLI Arguments
Because of its simplicity, this commit holds several changes:

1. Removal of some CLI Arguments

We decided to use recursive globs to be able to significantly simplify
our CLI arguments and settings.

2. Removal of unneeded dest keyword arguments

If argparser has an argument --some-thing it will automatically store
this thing in some_thing. There is no need to specify this manually
every time.

3. Reformatting