parsing: Move regex test patterns into setup
Regex tests now doesn't need to redeclare everytime the different
patterns. For that purpose the self.bs member moves to the top of the
setup because the regex-test-patterns need this variable.
Also add another expected value for the test
`test_unescaped_split_regex_pattern()` since the previous test pattern
here missed the last regex that is now in the class-wide test-pattern
list.