dbus: Add DbusApp
Add the DbusApp class - it stores information about evey application
that connects to coala via dbus. The DbusApp class is a container to
store data and has no functionality as such, hence there are no tests
for it. But this class will be used in other classes, and hence will be
used in their tests.