TestHelper: Remove paths added to sys.path
When running tests, we need to include the directory fo the test file
in the path so that it can be imported easily. But this causes issues
as th PATH becomes cluttered, and module names may clash.
So, we use the preserve_sys_path context manager.

Fixes https://github.com/coala-analyzer/coala/issues/574