Add ShowBears.py to coalib.output
ShowBears.py contains process_bears which extracts all the bears from
a section and compiles them in a dictionary. Then show_bears method
calls this for every section which is enabled or is present in the
targets and calls the interactor to print out the information.

See https://github.com/coala-analyzer/coala/issues/477