CloneDetectionRoutines: Prefer big code clones
This reduces the calculated difference value for pairs of bigger clones.
This makes sense as bigger function with the same relative difference
are more likely to provide a refactoring opportunity than small
functions. As a result of this change small functions are less likely to
be recognized as clones which is a good property mirroring the real
usefulness in theory. As the testsuite still passes this change does not
break existing testcases, as more are added, the usefulness of this
change will be evaluated.