circleci: Set default python version as 3.4.2
Earlier we could only have 1 container with python 3.4.2 as that was
checked when sending coverage information to codecov and coveralls.
Now, the condition has been changed to the 0th container in circleci
Hence, set the python version choices such that 3.4.2 is chosen as
the default version - and 3.3 and 3.2 are chosen for specific containers.