circleci: Clean up install scripts
In this commit, we set the quiet flag for pip - so that it only shows
erros and doesn't clutter the logs. In install scripts, redirect make to
/dev/null and rerun it if the command fails only. This is the recommended
way of silencing make.

Also, install scripts don't need to source setup_env_vars as it exports
the variables - and they are set in deps.sh already.