TestHelper: Run tests using multiple threads
This uses the `imap` function of multiprocessing.Pool to run multiple
tests simultaneously. Also, return the output so that it can be printed
in the correct order. Finally, increase some of the timeouts used in
various places - as some tests which create subprocesses within them
take more time to complete now - as it needs to wait for other processes
to free up before executing.