circleci: Use 3 jobs and reduce waiting time
In circleci, the get_cpu_count() gives a value of 24. And circleci
doesnt allow our container to use all of those. That is why each test
is queued for a long time in circleci.
Also, reduce timeouts and waiting times in tests as the circleci issue of
using too many jobs is resolved and each test shouldn't take too much
time now.