Interactor: Clear file_diff_dict on new section
Currently, if an interactor is reused for multiple sections, the file_diff_dict
is not cleared - and gives a KeyError in finalize when the file saved in
file_diff_dict is not found in the list of files for that section.
Here, we clear that dict whenever a new section is used - i.e. when
begin_section() is called.