coalib: Add coala_json
The coala-json file will be used to show a json output for coala. This
can be used to store the results for later use and has no interaction
with the user. It can also be used as an API for other programs, where
they simply execute `coala-json ...` and get back a easily parsable
output in json.

Fixes https://github.com/coala-analyzer/coala/issues/764