dbus: Change org.coala to org.coala_analyzer
As coala.org is an existing url, we cannot use it. Switch to the
url coala-analyzer.org as we own this url. Also, dbus standards don't
allow hyphen(-) to the object_path and interface names, so it becomes
`org.coala_analyzer`.

This was checked using a grep command.

Fixes https://github.com/coala-analyzer/coala/issues/822