Result: Use SourceRanges for code relations
This change finally enables bears to associate multiple parts in the
source code with one result. Of course this results in heavy changes in
a lot of other components that use the result. It also implies that the
console output needed to be reworked to now only show relevant lines
without context - otherwise e.g. overlapping regions get really
complicated. Showing context again is a task for another time, this
commit is big enough.