docs: Escape glob argument
In zsh, under default settings, when globs are specified by '=' in
arguments, the asterisk (*) is taken to be a literal asterisk.

So, in here, if the --files argument is --files=src/*.c, it looks
for a file named *.c in src/folder, and wouldn't act as a glob,
i.e. take in account all files in the src/ folder with a '.c'
extension.

So to avoid this, the asterisk is escaped with a leading forward
slash so that it acts as a glob.