DbusTest: Refactor to allow pytest execution
Refactor DbusTest so that it can be run by pytest. As pytest cannot
run the file again, we create a subprocess with a string for python to
execute to create the DbusServer.

Also, DbusServerTest was not closing the DBus server spawned by it.
This causes an issue where they both cannot be run by pytest
simultaneously. This commit kills the bus created in DbusServerTest
during tearDown so that it doesn't affect other tests.