Collectors: Add limit_paths() & expose it in args
There was no method to limit the files which coala would run
on in the command line arguments (without changing the coafile). This
commit adds a new argument namely "--limit-files" which ensures that
coala does not analyze any file outside the glob patterns mentioned in
this argument.

This is useful specially when coala's analysis is required in only 1
file like in the case of editor plugins.

Fixes https://github.com/coala-analyzer/coala/issues/1315