ConfWriter: Write back escape characters correctly
Earlier, escaped characters like the ones for comments
key_val delimiters, etc were not being written back
correctly. The value written back would be wrong because
the \ was not present anymore => the next time the config
file is read, it parses it differently.
This commit fixes it by taking these values and replacing
then with a backslash in front.

Fixes https://github.com/coala-analyzer/coala/issues/921