DefaultArgParser: Allow different argparsers
This commit allows coala to use multiple arg parsers
(with overlaps in their args). We add an extra argument
to default_arg_parser which chooses which entry_point
coala was run from. If it detects the entry point to be
coala-json, it will return a different argparser.

Fixes https://github.com/coala-analyzer/coala/issues/1375