coala-json: Add new arg --text-logs
coala-json normally returns the logs as JSON. Using
this arg, the user can select to see the logs in the normal
text format. This helps because the user doesnt have
to wait for the analysis to complete before seeing the
logs on their terminal.

Fixes https://github.com/coala-analyzer/coala/issues/847