ConfWriter: Take multiple delimiters but use one
ConfWriter needs all the delimiters being used in the coafile
because it will beed to escape chars appropriately. In this commit
we set which chars are being used as delimiters as iterables. But
we choose the first item in the iterable to be the value we will use
to write the conf writer - for uniformity.