coafile: Add PyImportSortBear and apply patches
The PyImportSortBear can take in a file and sort the imports
accordign to alphabetic order. This make it easy to get organized.
This commit adds the PyImportSortBear to the coafile and sets
the default action as AutoApplyPatch.
This allows isort to fix any file it finds to be wrong easily.

Fixes https://github.com/coala-analyzer/coala/issues/868