Lint: Make lint use stdout/err to create Diffs
Similar to the funtionality provided by CorrectionBasedBear,
now Lint can also take in stdin and create diffs+results.
You can use the `gives_corrected` value to tell Lint whether
the executable suggests the fix or not. If it does give the
fix, we parse the stdout/err to get the modified lines.