LocalBearTestHelper: Simplify code and names
Simplify the code and names of functions so that it is
more intuitive and easier to read.
This is the last of a series of commits to revamp the
LocalBearTesthelper class.

The code was repetitive. This makes some things common.
Also, this will lead to a better assert message because
the `lical_bear.execute()` function is outside the assert
so, it'd show the result in the Traceback also.

The names of functions weren't according to our codestyle.
Also, the names were too long to be able to make good git
shortlogs using them.

Fixes https://github.com/coala-analyzer/coala/issues/1370