ContextManagers: Add prepare_file()
Add context manager prepare_file() which can create a tempfile
and save the file contents there temporarily (for a bear).

The arg create_tempfile can be used to tell the Helper
to create a tempfile and give that filename to the bear.
The arg force_linebreaks adds a linebreak (\n) at the end of
each line if needed.