coalib: Use non-zero exitcode even if patches work
When coala has results it normally gives the exitcode 1.
But when the patches have results, and we autoapply the
patch, it seems to give 0 when there are no results after
the patches get applied. This is bad because the user won't
know that there were any results in the first place.
This commit sets the exitcode in such scenarios to 5.

Fixes https://github.com/coala-analyzer/coala/issues/1187