PyLintBear: Add new argument rcfile
Now that Lint shows warnings from stderr, pylint raises the
issue if no rcfile being found. This tells pylint to use
devnull as the rcfile if it cannot find anything else.
We also add an argument just in case bear users want to
specify the rcfile.
We also remove the default --disabled value "fixme" as it
conflicts with the rcfile if provided.

Fixes https://github.com/coala-analyzer/coala/issues/1457