coala_main: Move autoapply to gather_configs()
gather_configurations() is meant to gather all configurations.
autoapply is one such config but it has a special handle as it can
also be set from coala_main's args as well as CLI and coafile.
This moves the logic for this from coala_main to
ConfigurationGathering.

Related to https://github.com/coala-analyzer/coala/issues/1468