TextRange: Copy TextPosition by value
Earlier, end was the set to be the same as start. This meant
that if a TextRange was created with (TextPosition(1), None)
then: start is end == True.
This is bad because later if the range is extended, they both
cannot be modified independently. This commit uses a deepcopy
function to ensure this doesn't happen.

Fixes https://github.com/coala-analyzer/coala/issues/1069