LineDiff: Accept any iterable in add_after()
add_after() earlier only accepted lists. Now that we also use
tuples to handle these cases, we need it to accept any iterable.
We check if the lines to add after are greater than 1 if its an
iterable. If it is, we convert it into a list and save it. If not,
we use False.