ClangAstPrintBear: Convert to GlobalBear
The output of the bear was confusing as it was mixed up. This was
because LocalBears run on files, and the order in which the files
were being analyzed isn't the same between runs.
Now, we convert it to a global bear so that it run on files after
sorting them.

Fixes https://github.com/coala-analyzer/coala/issues/944