Shell.run_shell_command: Remove `shell=True`
This protects us against shell injection and also surprises due to
shell interpretation.

For compatability use `shlex.split()` when providing a string as a
command and not a sequence.

Also the `ShellTest` needs adaption since it used a shell-command
before which is not possible any more. Now argument-lists are used
instead.

Fixes https://github.com/coala-analyzer/coala/issues/1264.