TestUtilities: Add helpers to use test module
Notes:
I tried creating a setup.py for this module and installing it
using `import pip` and so on inside the unittest, but it failed.
This was because I needed to refresh pkg_resources, and on doing
`importlib.reload()` coverage used to throw errors.
Also, pip needs sudo in some computers and to overcome this
`--user` could be used - but this had it's own problems. i.e. the
USER pythonpath was not added to sys.path by default, which means
we had to `reload(site)` also. And even then pip couldn't uninstall
it if it was installed with --user.

Related to https://github.com/coala-analyzer/coala/issues/1610