ConsoleInteraction: Spaces -> Dots, Tabs -> Arrows
When coala reports SpaceConsistency for trailing spaces, there is
no way to understand where the trailing spaces are. This will
show the spaces and tabs as unicode dots and arrows respectively.

Spaces will be shown as •
Tabs (of width 4) will be shown as --->

Fixes https://github.com/coala-analyzer/coala/issues/606