Version: Change version to dev9999
When we use the git version and try installing things,
it gets messed up because the git has dev0 which is
lower than the current nightly release. This is
conceptually wrong as the latest thing in git is
the latest most beta version of coala.

Fixes https://github.com/coala-analyzer/coala/issues/1800