*.py: Use double backtick for code tag in RST
Earlier we used markdown and it had single backtick for the code
tag. Now RST silently fails for single backtick without a warning.
This commit changes all instances of single backtick to double
backtick to get back the original styling.

The command: git grep " \`[^\`]" coalib/
was used to list all uses of single backtick and I went through them
to find instances that were in docstrings.