coala: Catch exceptions when showing bear info
Earlier, coala_main.py had a broad exception to ensure no tracebacks
are shown. Now, we move it to coala.py so that coala-json, coala-format,
etc. are not affected by it. On doing this, a lot of errors were not
being caught in the broad exceptions making people see ugly tracebacks.

This commit fixes this by adding the try-catch appropriately and using
unittest.mock to test this with the exception being raised in a side
effect.

Fixes https://github.com/coala-analyzer/coala/issues/1977