Bear: Allow LANGUAGES to be a single string
It is inconvinient to keep writing LANGUAGES = ("java",).
Rather the alternative syntax to allow both a single string or
a tuple was proposed. This makes it simpler to write LANGUAGES for
bears with just 1 language.

This is also modified in ConsoleInteraction where it earlier expected
a iterable. Now, it changes it automatically to a tuple if it is a
string.

Fixes https://github.com/coala-analyzer/coala/issues/2068