coalaTest: Use bear_test_module() in coala -A test
When testing coala -A we do not want other bear modules which have
errors to make the test fail. Hence, we force it to only use
the `test_bears` module which we have.

This is done using the contextmanager bear_test_module() which overrides
the pkg_resources end-point finding utility.

Fixes https://github.com/coala-analyzer/coala/issues/2055