README: Use width = 100% for asciinema
It seems github strips the width information from the image tags
due to formatting/security concerns. They prefer making their
own max-width attribute and using that, which is set to 100%
in the HTML and cannot be configured in rst.

Hence, this commit attempts to fix pypi formatting troubles
using the :width: and hoping pypi does not ignore that
attribute.

Fixes https://github.com/coala-analyzer/coala/issues/2052