Linter: Omit partialmethod'ed signatures
If no explicit value is provided for a default value when using a
built-in output format, then it still appears in the partialmethod'ed
signature. `FunctionMetadata` grabs the metadata and forwards them to
the coala help. Now all arguments are omitted when defining a built-in
output format.

Fixes https://github.com/coala-analyzer/coala/issues/2102