DocumentationComment: Add assemble()
This function takes in a ``DocumentationComment`` instance and
assembles it to a documentation comment as a string.

To have tests on all corner cases, documentation styles in
default.py has changed.

Also, a default.c is created to test corner cases, for a docstyle
having a different marker. (i.e. marker[1] != "")