Replace or with inline if and vice-versa
Does not replace "or" in line 30 of
coalib/bearlib/languages/LanguageDefinition.py and
in line 139 of
coalib/bearlib/languages/documentation/DocstyleDefinition.py
because for any empty case of coalang_dir, we need the other one
and not the empty case.

Also does not replace if x is boolean.

Fixes https://github.com/coala-analyzer/coala/issues/1544