Collect bears only once across sections
Previously, bears were collected for each section in the configuration
file with no regard for other sections. This meant that the same bear
could be collected multiple times if it appeared in multiple sections.

Now, a list of unique bears to be collected is built from all the
sections before any collection takes place. This means that no
collection is repeated.

Fixes https://github.com/coala-analyzer/coala/issues/2075