LogPrinter: Change log type DEBUG to log type info
Instead of asking the user to run coala with -L DEBUG It's  better to have exception tracebacks default.

Closes https://github.com/coala/coala/issues/2660

ShowPatchAction: Shorten descriptions of actions

The users do not read the long descriptions.
It is, like this, easier for users to read
the options and not ignore them.

Closes https://github.com/coala/coala/issues/2701

result_actions: Correct docstring format

Updates docstring format for two functions
in the result_actions directory.

Closes https://github.com/coala/coala/issues/2779

circle.yml: Remove -V tag

The coala-ci -V command in circle.yml shows a lot of unnecessary info.
Removing it helps display only the relevant information.

Closes https://github.com/coala/coala/issues/2777