Requirements: Change 'manager' param to 'type'
This change is in line with the new approach of cib to wrap managers and
choose the right ones for a requirement. Managers should no longer be
determinded by the requirement, as i.e. a python requirement can be
installed by pip or conda, and a DistributionRequirement can be
installed by zypper, apt, pacman or whatever manager is available.