PackageRequirement: Rename to Requirement
This brings the naming scheme more in line with our usual style, i.e.
PEP8Bear(Bear) etc. It also avoids confusion, because PackageRequirement
can be interpreted as referring to "system packages".

This commit does not break compatibility with coala-bears at this point
in time, as no bear is using the PackageRequirement class directly.