Requirements: Change ``manager`` to ``type``
This change is in line with the new approach of ``cib``
to wrap amangers and choose the right ones for a requirement.
Managers should no longer be determined by the requirement,
as a ``python`` requirement can be installed by ``pip``
or ``conda``, and a DistributionRequirement can be installed
by ``zypper``, ``apt`` or whatever manager is available.