Caching: Store files to untrack and apply later
Say a file X and two sections A and B. If A yields a result
for X, the file is untracked. But before B executes, all files
are tracked again. So if B doesn't yield a result for X, the
file is wrongly thought as no-results for this run. Hence the
file will never be collected (at least until the next time the
file is changed). This is pretty critical.

Fixes https://github.com/coala/coala/issues/2987