Newcomers_Guide.rst: Specified the cobot assign <issue_link> command
Added a shor paragraph to Step 3 of the Newcomers_Guide in which it
is clearly specified what a newcomer should do in order to get an
issue. Tried to do it in a friendly and entertaining way. It was a
needed change because as it was in my case there were many newcomers
who forgot that they shouldn`t work unasigned on an issue and also
there were newcomers who didn`t know how to assign themselves to an
issue and there appeared a lot of not really necessary questions on
chat.

Fixes https://github.com/coala/coala/issues/3495