CliParsingTest.py: Use unittest.asserRaisesRegex
Using unittest.assertRaises should be avoided. unitest.assertRaisesRegex
has an extra parameter to check the exception message, and should be
used instead. By checking the exception message, the unit test verifies
that the exception is precisely the one which was expected, rather than
only of the same type as expected.

Related to https://github.com/coala/coala/issues/3374