Writing_Good_Commits.rst: Update Issue reference
When passing through the Issue reference paragraph of the
Writing_Good_Commits file it was not clear whether people
should use Fixes/Closes when referring to a docs 'bug'.
This creates confusion and unuseful discussions on the
pull requests. Thus, it is now more clear in which situations
you should use one or the other.

Closes https://github.com/coala/coala/issues/3280