Newcomers_Guide.rst: Specify the assign command
Added a short paragraph to Step 3 of the Newcomers_Guide in which it
is clearly specified what a newcomer should do in order to get an
issue. It was a needed change so that newcomers wouldn`t work
unasigned and also it would be pretty easy for them to find the way
to assign themselves to the respective issue.

Closes https://github.com/coala/coala/issues/3495