Useful_Links.rst: Add Useful_Links.rst
Especially for newcomers there was an issue about
their orientation through the site in terms of finding
specific materials that can help them. Adds a new page
with the most useful stuff from a newcomer`s perspective.
Edits the index file from the root folder so that the
document can be added to the page.

Closes https://github.com/coala/coala/issues/3515