WIP ConsoleInteraction: Show context for results
When results are displayed on console, diff and lines are printed.
To make it easier and understandable, this commit adds changes to
print context of these lines to console. Context is printed for
both affected lines and diff.
Tests have not been written. Once the changes are satisfactory and
approved, tests will be added.
Approach: In these changes, line numbers of necessary context are
passed to ShowPatchAction. Followed this approach, to avoid
recalculating these in ShowPatch.

See https://github.com/coala/coala/issues/2745