Add platform specific branch coverage pragma
Branch coverage defaults to requiring coverage of the `else`
statement for every `if` or `for` statement.
For a platform specific branch, the `else` branch will never
be visited.  Where there is an `else` statement, for other
platforms, the opposite platform specific coverage pragma
can be used to specify the branch doesnt need to be covered.
However there are a few cases where there is no `else` branch,
so a partial branch pragma is needed to indicate there is only
one possible branch for the platform.

By only creating platform specific partial branch pragma,
this change also disables the default `pragma: no branch`.

Related to https://github.com/coala/coala/issues/4166