Ignore Action: Remove action after comment added
This avoids asking the users if they want to add an ignore comment,
AFTER they already applied one.
Includes a new parameter for Action.is_applicable, which can be
checked for earlier actions.
Change in behaviour for IgnoreResultAction:
If this action was already used before, Action.is_applicable will
always return the appropriate comment.
Checks for "True" on the "apply_single" parameter to avoid infinite
loop.

Closes https://github.com/coala/coala/issues/3375