Processing: Allow Bears to use raw files
Processing allows raw files if the Bear state that it uses raw
files. Bear that uses binary files can't be mixed with Bear that uses
text files. If this raw files is enabled, the bears themselves are
in charge of doing the file handling (opening it, closing it,
reading it, etc).

Closes https://github.com/coala/coala/issues/3529