Diff.py: Add linebreaks to modified and original
This change adds linebreaks to the original and modified
properties, considering both the original lines as well as changes.
Linebreaks are only added, if they were not present before.
Especially the presence (or lack thereof) of a linebreak at EOF
will be preserved.
This rule is only broken, if the last line of the file is deleted.

Closes #3834