AspectList: Add ``exclude`` attribute
Extend AspectList class to have `exclude` attribute that hold list of
aspect that marked as exception of the list. Trying to access aspect
or subaspect of item in `exclude` will always give False/None.

```.python
>>> Metadata not in AspectList([Root], exclude=[Metadata])
True
>>> AspectList([Root], exclude=[Metadata]).get('metadata')
None
```

This feature will be used to excluding an aspect (to "deactivate") that
was defined in user's coala configuration.

Closes https://github.com/coala/coala/issues/4438