BearTest: Remove urllib3's ProtocolError
When requests was undergoing frequent breakages, a `ProtocolError`
was being raise unexpectedly when provoking a broken read scenario.
This exception was added as an alternative as it was at least
a predictable error occurring, allowing the remainder of the
test method to verify the resulting behaviour of the method,
and there were far bigger problems occurring at the time
due to `requests` being broken several time a day.

Depending on urllib3 for this ProtocolError is problematic,
as requests has its own version requirements for urllib3,
and those must take precedence over coala testing needs.
A conflict between requests and urllib3 has broken the
docker image.

It seems ProtocolError is not being thrown any longer,
and removing it bypasses the urllib3 conflict.
Should a ProtocolError occur again, it should be
investigated.

Fixes https://github.com/coala/coala/issues/4512