DocBaseClass.py: Move class to new file
Rename the existing DocBaseClass.py to DocumentExtraction.py
and move the DocBaseClass to a new DocBaseClass.py.

Closes https://github.com/coala/coala/issues/4507

DocBaseClass.py: Move class to new file

Rename the existing DocBaseClass.py to DocumentExtraction.py
and move the DocBaseClass to a new DocBaseClass.py.

Closes https://github.com/coala/coala/issues/4507

DocBaseClass.py: Move class to new file

Rename the existing DocBaseClass.py to DocumentExtraction.py
and move the DocBaseClass to a new DocBaseClass.py.

Closes https://github.com/coala/coala/issues/4507

DocBaseClass.py: Move class to new file

Rename the existing DocBaseClass.py to DocumentExtraction.py
and move the DocBaseClass to a new DocBaseClass.py.

Closes https://github.com/coala/coala/issues/4507