Language.py: Fix __contains__ method
The __contains__ method in the Language class only checks
if self.versions is a superset of item.versions. As a result
statements such as 'python' in Language['python 3'] returns False.
This change restructures the method so it will return True
if self.versions is a subset or superset of item.versions.

Closes https://github.com/coala/coala/issues/4690