.travis.yml: Skip unsupported on pull requests
The jobs for unsupported platforms 2.7 & 3.3 do not need to run
so frequently, and any breakages are usually related to upstream
dropping support for 2.7 or 3.3, rather than a local problem.
The master builds will break, and usually we'll need to pin
something to get around the problem.  Ultimately, failure to
verify that these old versions fail in a predictable manner
should not prevent us adding new features or dependencies.

Closes https://github.com/coala/coala/issues/4323