Development_Setup.rst: Rephrase instructions
Make the first sentence of "Installing from Git"
simple and comprehensible as it is unintelligible.

Remove the second code sample as it is identical
to the previous code sample in what it achieves,
but it is described as if it is different.

Append to the description of the first code sample
in what it achieves.

Closes https://github.com/coala/coala/issues/3903