Add decorator add_value_checks
The decorator may work on any function but made for the
specific use of the bears' `run()` methods or the
`create_arguments()` methods. This decorator compares the value
given to the bear setting against a list of acceptable values
provided by the bear writer. The decorator accepts a list of
tuples with each tuple containing two values, the bear setting
name as a string and a list of acceptable values.