DefaultArgParser: Restructure argcomplete support
Use a global symbol for the import, to avoid
multiple invocations of `import` and so that
the optional import can be tested, adding
coverage of that scenario using a mock import.

Also avoid the lambda in order to have coverage
of that line.

Delay loading of coalib.collecting.Collectors
as default_arg_parser is called from setup.py,
and coala-utils may not be installed yet.

Add test cases which exercise the possibilities
of unavailable imports.

Related to https://github.com/coala/coala/issues/4166
Related to https://github.com/coala/coala/issues/5664