Release Notes: Publish in api docs
Add release_notes in the index page of api doc.
Because RELEASE_NOTES.rst is located in root directory i.e,
../RELEASE_NOTES.rst, Solution is to create another
release_notes.rst file in docs dir that points to root .rst file.

Every release notes are in the single .rst file and all
have individual titles.
Due to this, it is rendering all release notes title in the
Home navigation. The solution is to create its own separate
toctree.

Closes https://github.com/coala/coala/issues/5724