coalaTest: Add test preventing unknown warnings
Currently the test_coala_with_color and _without_color
tests fail with little indication of what is happening,
especially as the first warnings messages are related to
the .coafile 'Default' section.  The reported error is
trimmed, preventing the real cause of the error from
being seen in the logs.  It can be hiding a backtrace.

This introduces a new test which explicitly checks for
the two types of warnings which are currently happening,
being the 'Default' section and missing bears warnings.
These are happening because the repository .coafile is
being loaded instead of only using CLI arguments.

Then any additional warnings are considered to be an
error, and the entire output is show, so the reader
can see what is causing the problem.

Related to https://github.com/coala/coala/issues/5390