NlSection.py: Add NlSection modules and it's tests
This PR adds the class NlSection.

In Nested Language mode, we divide the original file
into various sections. These sections are called as
`pure language sections` and `mixed language` sections.

NlSection helps us to define these sections and also
maitain the information of the start and the end position
of these sections in the original file. This information
helps us while linting.

This has an similar implementation to that of the SourceRange
of coala which is used to store the information about the
affected code.

This PR also has the tests for NlSection.